update Category domain namesapce
