+++ src/Core/Domain/Category/CommandHandler/ToggleCategoryStatusHandlerInterface.php
@@ -24,51 +24,17 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Category\ValueObject;
+namespace PrestaShop\PrestaShop\Core\Domain\Category\CommandHandler;
 
-use PrestaShop\PrestaShop\Core\Domain\Product\Category\Exception\CategoryException;
+use PrestaShop\PrestaShop\Core\Domain\Category\Command\ToggleCategoryStatusCommand;
 
 /**
- * Class CategoryId.
+ * Interface ToggleCategoryStatusHandlerInterface.
  */
-class CategoryId
+interface ToggleCategoryStatusHandlerInterface
 {
     /**
-     * @var int
+     * @param ToggleCategoryStatusCommand $command
      */
-    private $id;
-
-    /**
-     * @param int $id
-     *
-     * @throws CategoryException
-     */
-    public function __construct($id)
-    {
-        $this->setId($id);
-    }
-
-    /**
-     * @return int
-     */
-    public function getValue()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param int $id
-     *
-     * @throws CategoryException
-     */
-    private function setId($id)
-    {
-        if (!is_numeric($id) || 0 >= $id) {
-            throw new CategoryException(
-                sprintf('Invalid Category id %s supplied', var_export($id, true))
-            );
-        }
-
-        $this->id = (int) $id;
-    }
+    public function handle(ToggleCategoryStatusCommand $command);
 }
