+++ src/PrestaShopBundle/Resources/config/services/adapter/category.yml
@@ -23,10 +23,10 @@ services:
     class: PrestaShop\PrestaShop\Adapter\Category\CommandHandler\UpdateCategoriesStatusHandler
     tags:
       - name: tactician.handler
-        command: PrestaShop\PrestaShop\Core\Domain\Product\Category\Command\UpdateCategoriesStatusCommand
+        command: PrestaShop\PrestaShop\Core\Domain\Category\Command\UpdateCategoriesStatusCommand
 
   prestashop.adapter.category.command_handler.toggle_category_status_handler:
     class: PrestaShop\PrestaShop\Adapter\Category\CommandHandler\ToggleCategoryStatusHandler
     tags:
     - name: tactician.handler
-      command: PrestaShop\PrestaShop\Core\Domain\Product\Category\Command\ToggleCategoryStatusCommand
+      command: PrestaShop\PrestaShop\Core\Domain\Category\Command\ToggleCategoryStatusCommand
